% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentocorpus.R
\name{as.sento_corpus}
\alias{as.sento_corpus}
\title{Convert a quanteda or tm corpus object into a sento_corpus object}
\usage{
as.sento_corpus(x, dates, do.clean = FALSE, ...)
}
\arguments{
\item{x}{a quanteda \code{\link[quanteda]{corpus}} object, or a tm
\code{\link[tm]{SimpleCorpus}} or \code{\link[tm]{VCorpus}} object.}

\item{dates}{a sequence of dates as \code{"yyyy-mm-dd"}, of the same length as the number of documents
in the input \code{corpus}.}

\item{do.clean}{see \code{do.clean} argument from the \code{\link{sento_corpus}} function.}

\item{...}{...}
}
\value{
A \code{sento_corpus} object, as returned by the \code{\link{sento_corpus}} function.
}
\description{
Converts a \pkg{quanteda} \code{\link[quanteda]{corpus}} object into a \code{sento_corpus} object, by
adding user-supplied dates and doing proper rearranging.
}
\examples{
data("usnews", package = "sentometrics")
txt <- system.file("texts", "txt", package = "tm")
reut21578 <- system.file("texts", "crude", package = "tm")

# reshuffle usnews data.frame for use in quanteda and tm
dates <- usnews$date
# usnews$date <- NULL ### TODO: remove or not?
usnews$wrong <- "notNumeric"
colnames(usnews)[c(1, 3)] <- c("doc_id", "text")

# corpus conversion from a quanteda corpus
qcorp <- quanteda::corpus(usnews,
                          text_field = "text", docid_field = "doc_id")
corpus1 <- as.sento_corpus(qcorp, dates = dates)

# corpus conversion from a tm SimpleCorpus corpus (DataframeSource)
tmcorpSCdf <- tm::SimpleCorpus(tm::DataframeSource(usnews))
corpus2 <- as.sento_corpus(tmcorpSCdf)

# corpus conversion from a tm SimpleCorpus corpus (DirSource)
tmcorpSCdir <- tm::SimpleCorpus(tm::DirSource(txt))
corpus3 <- as.sento_corpus(tmcorpSCdir, dates = dates[1:length(tmcorpSCdir)])

# corpus conversion from a tm VCorpus corpus (DataframeSource)
tmcorpVdf <- tm::VCorpus(tm::DataframeSource(usnews))
corpus4 <- as.sento_corpus(tmcorpVdf)

# corpus conversion from a tm VCorpus corpus (DirSource)
tmcorpVdir <- tm::VCorpus(tm::DirSource(reut21578, mode = "binary"))
corpus5 <- as.sento_corpus(tmcorpVdir, dates = dates[1:length(tmcorpVdir)])

}
\seealso{
\code{\link[quanteda]{corpus}}, \code{\link{sento_corpus}}
}
\author{
Samuel Borms
}
