% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{betas}
\alias{betas}
\title{Compute beta weighting curves}
\usage{
betas(n, a = 1:4, b = 1:4)
}
\arguments{
\item{n}{a single \code{numeric} to indicate the lag length (cf., \emph{n}).}

\item{a}{a \code{numeric} as the first parameter (cf., \emph{a}).}

\item{b}{a \code{numeric} as the second parameter (cf., \emph{b}).}
}
\value{
A \code{data.frame} of beta weighting curves per combination of \code{a} and \code{b}.
}
\description{
Computes Beta weighting curves as in Ghysels et al. (2007). Handy to self-select specific time
aggregation weighting schemes for input in \code{\link{ctr_agg}}.
}
\details{
The Beta weighting abides by following formula:
\eqn{f(i/n; a, b) / \sum(i/n; a, b)}, where \eqn{i} is the lag index ordered from
1 to \eqn{n}, \eqn{a} and \eqn{b} are two decay parameters, and
\eqn{f(x; a, b) = (x^{a - 1}(1 - x)^{b - 1}T(a + b)) / (T(a)T(b))}{f(x; a, b)
 = (x^(a - 1) * (1 - x)^(b - 1) * T(a + b)) / (T(a) * T(b))}, where \eqn{T(.)} is
the \code{\link{gamma}} function.
}
\seealso{
\code{\link{ctr_agg}}
}
