% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{corpuS}
\alias{corpuS}
\title{Create a corpus}
\usage{
corpuS(text_df, minWords = NULL, maxWords = NULL)
}
\arguments{
\item{text_df}{a \code{data.frame} with as named columns: a document (\code{id}) column, a \code{date} column, a \code{text} column
(i.e. the columns where all texts to analyze reside), and a series of \code{feature} columns, with values pointing to the
applicability of a particular feature to a particular text. The latter columns are often binary (1 means the feature is applicable
to the document in the same row) or as a percentage to specify the degree of connectedness of a feature to a document. Features
could be for example topics (e.g. economic, political or legal), but also article sources (e.g. online or printed press), amongst
many more possibilities.}
}
\value{
A \code{corpuS} object.
}
\description{
Assembles a collection of texts (corpus), by calling the \code{corpus()} instructor from the \pkg{quanteda}
package and prepare it for further analysis.
}
