% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{ctr_agg}
\alias{ctr_agg}
\title{Setup control for aggregation into sentiment measures}
\usage{
ctr_agg(howWithin = "tf-idf", howDocs = "equal-weight",
  howTime = "equal-weight", do.ignoreZeros = FALSE, by = "day", lag = 1,
  alphas = seq(0.1, 0.5, by = 0.1), orders = 1:3, do.inverse = TRUE,
  do.normalize = TRUE, weights = NULL)
}
\arguments{
\item{howWithin}{a single character vector defining how aggregation within documents will be performed.}

\item{howDocs}{a single character vector defining how aggregation across documents per date will be performed.}

\item{howTime}{a single character vector defining how aggregation across dates will be performed.}

\item{do.ignoreZeros}{a logical indicating whether zero sentiment values have to be ignored while aggregation
across documents.}

\item{by}{a single character vector, either \code{"day", "week", "month"} or \code{"year"}, to indicate at what
level the dates should be aggregated. Dates will be displayed as the first day of the period, if applicable (e.g.
\code{"2017-03-01"} for March 2017).}

\item{lag}{a single integer vector, being the time lag to be specified for aggregation across time.}

\item{...}{other parameters required for the time weighting schemes. These are \code{alphas} if
\code{howTime == "exponential"}, and \code{orders}, \code{do.inverse} and \code{do.normalize} if \code{howTime == "almon"}.
One could also supply an own weighting scheme as a \code{data.frame} with the number of rows equal to the desired (implicit)
\code{lag}, via the \code{weights} option and for \code{howTime == "own"}.}
}
\value{
A list encapsulating the control parameters.
}
\description{
Sets up control for aggregation of document-level textual sentiment into textual
sentiment measures (indices).
}
\details{
For currently available options on how aggregation can occer (via the \code{howWithin},
\code{howDocs} and \code{howTime} parameters), call \code{get_hows()}.
}
\seealso{
\code{\link{get_hows}}
}
