% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{ctr_merge}
\alias{ctr_merge}
\title{Setup control for merging sentiment measures}
\usage{
ctr_merge(sentomeasures, lex = NA, feat = NA, time = NA,
  do.keep = FALSE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{lex}{a list with unique lexicons to merge at given name, e.g. \code{list(lex12 = c("lex1", "lex2"))}.}

\item{feat}{a list with unique features to merge at given name, e.g. \code{list(feat12 = c("feat1", "feat2"))}.}

\item{time}{a list with unique time weighting schemes to merge at given name, e.g. \code{list(tw12 = c("tw1", "tw2"))}.}

\item{do.keep}{a logical indicating if the original sentiment measures should be kept (i.e. the merged sentiment measures
will be added to the current sentiment measures as additional indices if \code{TRUE}).}
}
\value{
A list encapsulating the control parameters.
}
\description{
Sets up control for the optional merging (additional aggregation) of sentiment measures.
}
