% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmeasures.R
\name{ctrl_agg}
\alias{ctrl_agg}
\title{Setup control for aggregation into sentiment measures}
\usage{
ctrl_agg(howWithin, howDocs, howTime, ignoreZeros, lag, ...)
}
\arguments{
\item{howWithin}{a single character vector defining how aggregation within documents will be performed.}

\item{howDocs}{a single character vector defining how aggregation across documents per date will be performed.}

\item{howTime}{a single character vector defining how aggregation across dates will be performed.}

\item{ignoreZeros}{a logical indicating whether zero sentiment values have to be ignored while aggregation
across documents.}

\item{lag}{the time lag to be specified for aggregation across time.}

\item{...}{other parameters required for the time weighting schemes.}
}
\value{
A list encapsulating the control parameters.
}
\description{
Sets up control for aggregation of document-level textual sentiment into textual
sentiment measures (indices).
}
\details{
For currently available options on how aggregation can occer, call \code{get_hows()}.
}
\seealso{
\code{\link{get_hows}}
}
