% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{data-defunct}
\alias{data-defunct}
\alias{lexicons}
\alias{valence}
\title{Datasets with defunct names}
\description{
These are datasets that have been renamed and the old names removed. Please change your code to use the new names.
}
\details{
The dataset \code{lexicons} is defunct, use \code{list_lexicons} instead.

The dataset \code{valence} is defunct, use \code{list_valence_shifters} instead.
}
\keyword{internal}
