% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsoc2019.R
\name{.compute_sentiment.SimpleCorpus}
\alias{.compute_sentiment.SimpleCorpus}
\title{Only language available in metadata so no transformation to sentocorpus}
\usage{
.compute_sentiment.SimpleCorpus(x, lexicons, how, tokens = NULL,
  nCore = 1)
}
\description{
Only language available in metadata so no transformation to sentocorpus
}
