% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{get_dates}
\alias{get_dates}
\title{Get the dates of the sentiment measures/time series}
\usage{
get_dates(sentomeasures)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}
}
\value{
The \code{"date"} column in \code{sentomeasures[["measures"]]} as a \code{character} vector.
}
\description{
Convenience function. Returns the dates of the sentiment time series.
}
\author{
Samuel Borms
}
