% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get the dimensions of the sentiment measures}
\usage{
get_dimensions(sentomeasures)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}
}
\value{
The \code{"features"}, \code{"lexicons"} and \code{"time"} elements in \code{sentomeasures}.
}
\description{
Convenience function. Returns the components across all three dimensions of the sentiment measures.
}
\author{
Samuel Borms
}
