% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hows}
\alias{get_hows}
\title{Options supported to perform aggregation into sentiment measures}
\usage{
get_hows()
}
\value{
A list with the supported aggregation hows for arguments \code{howWithin} (\code{"words"}), \code{howDows}
(\code{"docs"}) and \code{howTime} (\code{"time"}), to be supplied to \code{\link{ctr_agg}}.
}
\description{
Outputs the supported aggregation arguments. Call for information purposes only. Used within
\code{\link{ctr_agg}} to check if supplied aggregation hows are supported.
}
\details{
Weighting within documents or sentences (\code{"words"}):
\describe{
\item{\code{"proportional"}}{divides each sentiment score by the total number of words.}
\item{\code{"proportionalPol"}}{divides each sentiment score by the number of detected
polarized words (counting words that appear multiple times by their frequency).}
\item{\code{"counts"}}{no normalisation.}
\item{\code{"squareRootCounts"}}{divides the sentiment by the square root of the number of tokens in each text.}
\item{\code{"UShaped"}}{gives a higher weight to words at the beginning and end of the texts.}
\item{\code{"invertedUShaped"}}{gives a lower weight to words at the beginning and the end of the texts.}
\item{\code{"exponential"}}{gives gradually more weight the later the word appears in the text.}
\item{\code{"invertedExponential"}}{gives gradually less weight the later the words appears in the text.}
\item{\code{"TF"}}{gives a weight proportional to the number of times a word appears in a text.}
\item{\code{"logarithmicTF"}}{gives the same weight as \code{"TF"} but logarithmically scaled.}
\item{\code{"augmentedTF"}}{weight is determined by dividing the raw frequency of a token by the raw frequency
of the most occurring term in the document (can be used to prevent a bias towards longer documents).}
\item{\code{"IDF"}}{uses the logarithm of the division of the raw frequency of a word by the number of
texts in which the word appears (words appearing in multiple texts get thus a lower weight).}
\item{\code{"TFIDF"}}{same weights as \code{"TF"}-variant but multiplied with \code{"IDF"} weights.}
\item{\code{"logarithmicTFIDF"}}{same weights as \code{"TF"}-variant but multiplied with \code{"IDF"} weights.}
\item{\code{"augmentedTFIDF"}}{same weights as \code{"TF"}-variant but multiplied with \code{"IDF"} weights.}
}

See the package's \href{https://ssrn.com/abstract=3067734}{vignette} for a detailed explanation of all
aggregation options.
}
\seealso{
\code{\link{ctr_agg}}
}
