% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmodel.R
\name{lm_sent}
\alias{lm_sent}
\title{Optimized sparse linear regression}
\usage{
lm_sent(sentmeasures, y, x, ctrlModel)
}
\arguments{
\item{sentmeasures}{a \code{sentmeasures} object.}

\item{y}{one-column \code{data.frame} of the dependent (response) variable.}

\item{x}{a named \code{data.frame} with other explanatory variables.}

\item{ctrlModel}{output from a \code{ctrl_model()} call.}
}
\value{
A list containing:
\item{reg}{optimized regression object.}
\item{alpha}{optimized calibrated alphas(s).}
\item{lambda}{optimized calibrated lambda(s).}
}
\description{
Linear model of a dependent variable on the wide number of sentiment measures and potentially
other explanatory variables. Either performs a linear regression on the provided variables at once, or computes
linear models sequentially for a given sample size over a longer time horizon, with associated forecasting performance
metrics.
}
\seealso{
\code{\link{ctrl_model}}, \code{\link[glmnet]{glmnet}}, \code{\link[caret]{train}}
}
