% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_measures_xyz.R
\name{measures_update}
\alias{measures_update}
\title{Update sentiment measures}
\usage{
measures_update(sentomeasures, sentocorpus, lexicons)
}
\arguments{
\item{sentomeasures}{\code{sentomeasures} object created with \code{\link{sento_measures}}}

\item{sentocorpus}{a \code{sentocorpus} object created with \code{\link{sento_corpus}}.}

\item{lexicons}{a \code{sentolexicons} object created with \code{\link{sento_lexicons}}.}
}
\value{
An updated \code{sentomeasures} object.
}
\description{
Updates a sentomeasures object based on a new corpus provided. Sentiment for the unseen corpus
texts calculated and aggregated applying the control variables from the input sentomeasures object.
}
\examples{
data("usnews", package = "sentometrics")

corpus1 <- sento_corpus(usnews[1:500,])
ctr <- ctr_agg(howTime = "linear", by = "year", lag = 3)
l <- sento_lexicons(list_lexicons[c("LM_en", "HENRY_en")],
                    list_valence_shifters[["en"]])
sentomeasures <- sento_measures(corpus1, l, ctr)
corpus2 <- sento_corpus(usnews[400:2000,])
sentomeasuresNew <- measures_update(sentomeasures, corpus2, l)

}
\seealso{
\code{\link{sento_measures}}, \code{\link{compute_sentiment}}
}
\author{
Jeroen Van Pelt, Samuel Borms, Andres Algaba
}
