% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmeasures.R
\name{merge_measures}
\alias{merge_measures}
\title{Merge sentiment measures}
\usage{
merge_measures(ctrlMerge)
}
\arguments{
\item{ctrlMerge}{output from a \code{ctrl_merge()} call.}
}
\value{
A modified \code{sentmeasures} object.
}
\description{
Merge (further aggregate) measures by combining across the lexicons, features and time weighting schemes
dimensions. The combination occurs by taking the mean of the relevant measures.
}
\seealso{
\code{\link{ctrl_merge}}
}
