% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{merge_to_global}
\alias{merge_to_global}
\title{Merge sentiment measures into one global sentiment measure}
\usage{
merge_to_global(sentomeasures, lex = 1, feat = 1, time = 1)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{lex}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$lexicons)}, in the same order and
summing to one. By default set to 1, which means equally weighted.}

\item{feat}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$features)}, in the same order and
summing to one. By default set to 1, which means equally weighted.}

\item{time}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$time)}, in the same order and summing
to one. By default set to 1, which means equally weighted.}
}
\value{
A modified \code{sentomeasures} object, with only the global sentiment measure, including updated statistics, but
the other list elements and the original sentiment scores \code{data.table} untouched.
}
\description{
Merges all sentiment measures into one global textual sentiment measure based on a set of weights provided to
indicate the importance of each component in the \code{lexicons}, \code{features} and \code{time} vectors as part of the
\code{sentomeasures} object. The global measure is composed as the multiplication of the individual weights across the
three dimensions times the sentiment value per date observation.
}
\seealso{
\code{\link{ctr_merge}}
}
