% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{nmeasures}
\alias{nmeasures}
\title{Get number of sentiment measures}
\usage{
nmeasures(sentomeasures)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}
}
\value{
The number of sentiment measures in the input \code{sentomeasures} object.
}
\description{
Returns the number of sentiment measures.
}
\author{
Samuel Borms
}
