% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{nobs}
\alias{nobs}
\title{Get number of observations in the sentiment measures}
\usage{
nobs(sentomeasures)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}
}
\value{
The number of rows (observations/data points) in \code{sentomeasures[["measures"]]}.
}
\description{
Returns the number of data points available for the sentiment measures.
}
\author{
Samuel Borms
}
