% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{nobs.sento_measures}
\alias{nobs.sento_measures}
\title{Get number of observations in the sentiment measures}
\usage{
\method{nobs}{sento_measures}(object, ...)
}
\arguments{
\item{object}{a \code{sento_measures} object created using \code{\link{sento_measures}}.}

\item{...}{not used.}
}
\value{
The number of rows (observations/data points) in \code{object[["measures"]]}.
}
\description{
Returns the number of data points available in the sentiment measures.
}
\author{
Samuel Borms
}
