% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{nobs.sentomeasures}
\alias{nobs.sentomeasures}
\title{Get number of observations in the sentiment measures}
\usage{
\method{nobs}{sentomeasures}(object, ...)
}
\arguments{
\item{object}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}

\item{...}{not used.}
}
\value{
The number of rows (observations/data points) in \code{sentomeasures[["measures"]]}.
}
\description{
Returns the number of data points available in the sentiment measures.
}
\author{
Samuel Borms
}
\keyword{internal}
