% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{perform_agg}
\alias{perform_agg}
\title{Aggregate textual sentiment across documents and time}
\usage{
perform_agg(toAgg, ctr)
}
\arguments{
\item{toAgg}{output from a \code{compute_sentiment()} call, a list with as main component a sentiment scores
\code{data.table} with dates and features--lexicons sentiment scores columns.}

\item{ctr}{output from a \code{ctr_agg()} call.}
}
\value{
A list containing:
\item{measures}{a \code{sentomeasures} object.}
\item{features}{a \code{character} vector of the different features.}
\item{lexicons}{a \code{character} vector of the different lexicons used.}
\item{time}{a \code{character} vector of the different time weighting schemes used.}
}
\description{
Condense document-level textual sentiment scores into a panel of textual sentiment
measures by aggregating across documents and time.
}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{ctr_agg}}
}
