% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{perform_MCS}
\alias{perform_MCS}
\title{Apply the model confidence set (MCS) procedure to a selection of models}
\usage{
perform_MCS(models, loss = c("DA", "errorSq", "AD", "accuracy"), ...)
}
\arguments{
\item{models}{a named list of \code{sentomodeliter} objects. All models should be of the same family, being either
\code{"gaussian"}, \code{"binomial"} or \code{"multinomial"}, and have performance data of the same dimensions.}

\item{loss}{a single \code{character} vector, either \code{"DA"} (directional \emph{in}accuracy), \code{"errorSq"}
(squared errors), \code{"AD"} (absolute errors) or \code{"accuracy"} (\emph{in}accurate class predictions). This argument
defines on what basis the model confidence set is calculated. The first three options are available for \code{"gaussian"}
models, the last option applies only to \code{"binomial"} or \code{"multinomial"} models.}

\item{...}{other parameters that can be supplied to the \code{MCS::MCSprocedure()} function.}
}
\value{
An object as returned from the \code{MCS::MCSprocedure()} function.
}
\description{
Calculates the model confidence set (see ``The Model Confidence Set"; Hansen, Lunde and Nason, 2011) as
implemented in the \pkg{MCS} package, for a set of different \code{sentomodeliter} objects.
}
\seealso{
\code{\link{sento_model}}, \code{\link[MCS]{MCSprocedure}}
}
\author{
Samuel Borms
}
