% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{predict.sentomodel}
\alias{predict.sentomodel}
\title{Make predictions from a sentomodel object}
\usage{
\method{predict}{sentomodel}(object, newx, type, offset = NULL, ...)
}
\arguments{
\item{object}{a \code{sentomodel} object.}

\item{newx}{a \code{matrix} of \code{numeric} values for all explanatory variables at which predictions are to be made,
see documentation for \code{\link{predict.glmnet}}.}

\item{type}{type of prediction required, a value from \code{c("link", "response", "class")}, see documentation for
\code{\link{predict.glmnet}}.}

\item{offset}{values to use as offset, only if an offset was also used in the model fitting, see documentation for
\code{\link{predict.glmnet}}.}

\item{...}{not used.}
}
\value{
A prediction output depending on the \code{type} argument provided.
}
\description{
Prediction (forecasting) method for \code{sentomodel} class, with usage along the lines of \code{predict.glmnet}, but
simplified in terms of allowed parameters.
}
\seealso{
\code{\link{predict.glmnet}}, \code{\link{sento_model}}
}
