% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{retrieve_attribution}
\alias{retrieve_attribution}
\title{Retrieves top-down sentiment attribution given forecasting equation}
\usage{
retrieve_attribution(sentomodel, rows = NULL, ...)
}
\arguments{
\item{sentomodel}{a \code{sentomodel} object.}

\item{rows}{row indices (i.e. dates) for which the attribution needs to be calculated, based on the measures in the
\code{sentomeasures} object packed within the \code{sentomodel} input. By default \code{NULL}, which means attribution
is calculated for all rows.}

\item{...}{to do.}
}
\value{
A list with all dimensions for which aggregation is computed, as \code{data.table}s with a \code{"date"} and
\code{"attribution"} column for the rows indicated.
}
\description{
xxx
}
