% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{retrieve_attributions}
\alias{retrieve_attributions}
\title{Retrieves top-down sentiment attribution given forecasting model}
\usage{
retrieve_attributions(sentomodel)
}
\arguments{
\item{sentomodel}{a \code{sentomodel} object.}
}
\value{
A list with all possible dimensions for which aggregation can be computed, as \code{data.table}s with an
\code{"id"} (at document-level) column, a \code{"date"} column and a \code{"attribution"} column. ### TODO
}
\description{
### TODO
}
\seealso{
\code{\link{sento_model}}
}
