% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmeasures.R
\name{select_measures}
\alias{select_measures}
\title{Select a subset of sentiment measures}
\usage{
select_measures(sentmeasures, toSelect)
}
\arguments{
\item{sentmeasures}{a \code{sentmeasures} object.}

\item{toSelect}{a vector of components (lexicon, time weighting and features) which form the measures selected.}
}
\value{
A modified \code{sentmeasures} object, with only the measures required.
}
\description{
Selects a subset of sentiment measures.
}
