% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmeasures.R
\name{sentmeasures}
\alias{sentmeasures}
\title{One-way towards sentmeasures object}
\usage{
sentmeasures(corpus, lexicons, ctrlAgg)
}
\value{
A \code{sentmeasures} object.
}
\description{
Wrapper function which assembles calls to \code{compute_sentiment()} and \code{perform_agg()}, and includes the input
\code{corpuS} and computed sentiment scores in its output. Serves as the most direct way towards a panel of textual sentiment measures,
and a \code{sentmeasures} object.
}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{perform_agg}}
}
