% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentoapp.R
\name{sento_app}
\alias{sento_app}
\title{Shiny App}
\usage{
sento_app()
}
\description{
Shiny app to visually set-up sentiment computation.
}
\details{
The Shiny app is a visual wrapper around the \code{compute_sentiment} function. You can upload texts or a corpus in csv
format, choose built-in lexicons/valence shifters or upload your own lexicons/valence shifters in csv format. All the available
weighting schemas that are available in the sentometrics package can be selected and with one click on a button, you can calculate
the sentiment. Once calculated, the sentiment can be calculated as a CSV file. Additionally, there is some functionality coming
from the \code{corpus_summarize} function to gain insights in the corpus that you have uploaded.
}
\author{
Jeroen Van Pelt, Samuel Borms, Andres Algaba
}
