% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{sento_lm}
\alias{sento_lm}
\title{Optimized sparse linear regression}
\usage{
sento_lm(sentomeasures, y, x, ctr)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{y}{one-column \code{data.frame} of the dependent (response) variable.}

\item{x}{a named \code{data.frame} with other explanatory variables.}

\item{ctr}{output from a \code{ctr_model()} call.}
}
\value{
A list containing:
\item{reg}{optimized regression object.}
\item{alpha}{optimized calibrated alphas(s).}
\item{lambda}{optimized calibrated lambda(s).}
}
\description{
Linear model of a dependent variable on the wide number of sentiment measures and potentially
other explanatory variables. Either performs a linear regression on the provided variables at once, or computes
linear models sequentially for a given sample size over a longer time horizon, with associated forecasting performance
metrics.
}
\seealso{
\code{\link{ctr_model}}, \code{\link[glmnet]{glmnet}}, \code{\link[caret]{train}}
}
