% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{sento_model}
\alias{sento_model}
\title{Optimized sparse linear regression}
\usage{
sento_model(sentomeasures, y, x = NULL, ctr)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{y}{a one-column \code{data.frame} or \code{numeric} vector capturing the dependent (response) variable.}

\item{x}{a named \code{data.frame} with other explanatory variables, by default set to \code{NULL}.}

\item{ctr}{output from a \code{ctr_model()} call.}
}
\value{
If \code{ctr$do.iter == FALSE}, a list containing:
\item{reg}{optimized regression object.}
\item{alpha}{optimized calibrated alpha.}
\item{lambda}{optimized calibrated lambda.}
\item{trained}{output from \code{caret::train} call within function (if \code{ctr$type ==} "\code{cv}").}
\item{ic}{a \code{numeric} vector of minimum information criterion values per element of \code{alphas} (if \code{ctr$type !=} "\code{cv}").}
\item{type}{information criterion used to calibrate (if \code{ctr$type !=} "\code{cv}").}

If \code{ctr$do.iter == TRUE}, a list containing:
\item{regs}{optimized regression objects, i.e. lists as above.}
\item{alphas}{optimized calibrated alphass.}
\item{lambdas}{optimized calibrated lambdas.}
\item{performance}{a \code{data.frame} with performance-related measures, being "\code{RMSFE}" (root mean squared
forecasting error), "\code{MAD}" (mean absolute deviation), "\code{MAPE}" (mean absolute percentage error),
"\code{DA}" (directional accuracy), and each's respective individual values in the sample.}
}
\description{
Linear model of a dependent variable on the wide number of sentiment measures and potentially
other explanatory variables. Either performs a linear regression on the provided variables at once, or computes
linear models sequentially for a given sample size over a longer time horizon, with associated forecasting performance
metrics.
}
\seealso{
\code{\link{ctr_model}}, \code{\link[glmnet]{glmnet}}, \code{\link[caret]{train}}
}
