% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sentometrics-defunct}
\alias{sentometrics-defunct}
\alias{ctr_merge}
\alias{perform_MCS}
\title{Defunct functions}
\usage{
ctr_merge(...)

perform_MCS(...)
}
\arguments{
\item{...}{arguments not documented since function declared defunct.}
}
\description{
Functions defunct due to changed naming or because functionality is discarded. See the NEWS file for more information
about since when or why functions have been defunct.
}
\keyword{internal}
