% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sentometrics-deprecated}
\alias{sentometrics-deprecated}
\alias{fill_measures}
\alias{merge_measures}
\alias{subset_measures}
\alias{select_measures}
\title{Deprecated functions}
\usage{
fill_measures(sentomeasures, fill = "zero")

merge_measures(ctr)

subset_measures(sentomeasures, subset)

select_measures(sentomeasures, toSelect, do.combine = TRUE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}

\item{fill}{an element of \code{c("zero", "latest", NA)}; the first and last assume missing dates represent zero sentiment,
the second assumes missing dates represent constant sentiment.}

\item{ctr}{output from a \code{\link{ctr_merge}} call.}

\item{subset}{a logical expression indicating the rows to keep.}

\item{toSelect}{a \code{character} vector of the lexicon, feature and time weighting scheme names, to indicate which
measures need to be selected. One can also supply a \code{list} of such \code{character} vectors, in which case
\code{do.combine = TRUE} is set automatically, such that the separately specified combinations are selected.}

\item{do.combine}{a \code{logical} indicating if only measures for which all (\code{do.combine = TRUE}) or at least one
(\code{do.combine = FALSE}) of the selection components should occur in each sentiment measure's name in the selection. If
\code{do.combine = TRUE}, the \code{toSelect} argument can only consist of one lexicon, one feature, and one time weighting
scheme at maximum.}
}
\description{
Functions deprecated due to changed naming or because functionality is discarded. Deprecated functions are made defunct
every 1 major or every 2 minor package updates. See the NEWS file for more information about since when or why functions
have been deprecated.
}
\seealso{
\code{\link{measures_fill}}

\code{\link{measures_merge}}

\code{\link{measures_subset}}

\code{\link{measures_select}}
}
