% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentometrics.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Monthly S&P 500 Index returns}
\format{A \code{data.frame} with 322 rows and 2 columns.}
\source{
\href{https://finance.yahoo.com/quote/\%5EGSPC/history?p=\%5EGSPC}{S&P 500 (^GSPC) at Yahoo Finance}
}
\usage{
data("sp500")
}
\description{
Monthly returns for the S&P 500 Index between March 1988 and December 2014.

\itemize{
  \item date. Date as \code{"yyyy-mm-01"}.
  \item return. A \code{numeric} monthly return value.
}
}
\keyword{datasets}
