% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{summary.sentomodel}
\alias{summary.sentomodel}
\title{Summary of sentomodel object}
\usage{
\method{summary}{sentomodel}(object, ...)
}
\arguments{
\item{object}{a \code{sentomodel} object.}

\item{...}{not used.}
}
\value{
A sequence of informative prints on the model's results.
}
\description{
Prints out a short summary consisting of the main elements of the constructed model (model type, calibrated values, non-zero
coefficients and performance).
}
