% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentometrics.R
\docType{data}
\name{usnews}
\alias{usnews}
\title{Texts relevant (and not) to the US economy}
\format{A \code{data.frame}, formatted as required to be an input for \code{\link{sento_corpus}}.}
\source{
\href{https://www.crowdflower.com/data-for-everyone/}{Economic News Article Tone and Relevance}
}
\usage{
data("usnews")
}
\description{
A collection of texts annotated by humans in terms of relevance to the US economoy or not. The texts come from two major
journals in the US (The Wall Street Journal and The Washington Post) and cover 6801 documents between 1980 and 2014. It
contains following information:

\itemize{
  \item id. a \code{character} ID identifier.
  \item date. Date as \code{"yyyy-mm-dd"}.
  \item text. Texts in \code{character} format.
  \item wsj. Equals 1 if the article comes from The Wall Street Journal.
  \item wapo. Equals 1 if the article comes from The Washington Post.
  \item economy. Equals 1 if the article is relevant to the US economy.
  \item noneconomy. Equals 1 if the article is not relevant to the US economy.
}
}
\examples{
data("usnews")
usnews[3192, "text"]
usnews[1:5, c("id", "date", "text")]

}
\keyword{datasets}
