% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus.R
\name{add_features}
\alias{add_features}
\alias{add_features.corpuS}
\title{Add feature columns to corpuS}
\usage{
add_features(corpuS, features)

\method{add_features}{corpuS}(corpuS, features)
}
\arguments{
\item{corpuS}{a \code{corpuS} object.}

\item{features}{a named \code{data.frame} with as columns the new features to add to the \code{corpuS} inputted.}
}
\value{
An updated \code{corpuS} object.
}
\description{
Adds new named feature columns to provided \code{corpuS} object.
}
