% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{almons}
\alias{almons}
\title{Compute Almon polynomials}
\usage{
almons(n, orders = 1:3, do.inverse = TRUE, do.normalize = TRUE)
}
\arguments{
\item{n}{a single \code{numeric} to indicate the length of the curve (the number of lags, cf. \emph{n} in the formula).}

\item{orders}{a \code{numeric} vector as the sequence the Almon orders (cf. \emph{b} in the formula).}

\item{do.inverse}{\code{TRUE} if the inverse Almon polynomials should be calculated as well.}

\item{do.normalize}{\code{TRUE} if polynomials should be normalized to unity.}
}
\value{
A \code{data.frame} of all Almon polynomial weighting curves, of size \code{length(orders)} (times two if
\code{do.inverse == TRUE}).
}
\description{
Computes Almon polynomial weighting curves; handy to self-select specific time aggregation weighting schemes.
}
\details{
The Almon polynomial formula implemented is:
\eqn{(1 - (i/n)^{b})(i/n)^{max(b) - b}}{(1 - (i/n)^b) * (i/n)^(max(b) - b)}.
}
\seealso{
\code{\link{ctr_agg}}
}
