% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{ctr_agg}
\alias{ctr_agg}
\title{Setup control for aggregation into sentiment measures}
\usage{
ctr_agg(howWithin = "equal_weight", howDocs = "equal_weight",
  howTime = "equal_weight", do.ignoreZeros = FALSE, by = "day", lag = 1,
  alphasExp = seq(0.1, 0.5, by = 0.1), ordersAlm = 1:3,
  do.inverseAlm = TRUE, do.normalizeAlm = TRUE, weights = NULL)
}
\arguments{
\item{howWithin}{a single \code{character} vector defining how aggregation within documents will be performed. Should
\code{length(howWithin) > 1}, the first element is used.}

\item{howDocs}{a single \code{character} vector defining how aggregation across documents per date will be performed.
Should \code{length(howDocs) > 1}, the first element is used.}

\item{howTime}{a \code{character} vector defining how aggregation across dates will be performed. More than one choice
is possible here.}

\item{do.ignoreZeros}{a \code{logical} indicating whether zero sentiment values have to be ignored while aggregation
across documents.}

\item{by}{a single \code{character} vector, either \code{"day", "week", "month"} or \code{"year"}, to indicate at what
level the dates should be aggregated. Dates will be displayed as the first day of the period, if applicable (e.g.
\code{"2017-03-01"} for March 2017).}

\item{lag}{a single \code{integer} vector, being the time lag to be specified for aggregation across time. By default
equal to 1, meaning no aggregation across time.}

\item{alphasExp}{a \code{numeric} vector of all exponential smoothing factors to calculate weights for, used if
\code{"exponential" \%in\% howTime}. Values should be betwoon 0 and 1 (both excluded).}

\item{ordersAlm}{a \code{numeric} vector of all Almon polynomial orders to calcalute weights for, used if
\code{"almon" \%in\% howTime}.}

\item{do.inverseAlm}{a \code{logical} indicating if for every Almon polynomial its inverse has to be calculated too, used
if \code{"almon" \%in\% howTime}.}

\item{do.normalizeAlm}{a \code{logical} indicating if every Almon polynomial weights column should sum to one, used if
\code{"almon" \%in\% howTime}.}

\item{weights}{an own weighting scheme as a \code{data.frame} with the number of rows equal to the desired \code{lag}, used
if \code{"own" \%in\% howTime}.}
}
\value{
A list encapsulating the control parameters.
}
\description{
Sets up control object for aggregation of document-level textual sentiment into textual
sentiment measures (indices).
}
\details{
For currently available options on how aggregation can occer (via the \code{howWithin},
\code{howDocs} and \code{howTime} parameters), call \code{get_hows()}.
}
\examples{
# simple control function (using equal_weight default options)
ctr1 <- ctr_agg(howTime = c("linear"), by = "year", lag = 3)

# more elaborate control function (particular attention to time weighting schemes)
ctr2 <- ctr_agg(howWithin = "tf-idf",
                howDocs = "proportional",
                howTime = c("equal_weight", "linear", "almon", "exponential", "own"),
                do.ignoreZeros = TRUE,
                by = "day",
                lag = 20,
                ordersAlm = 1:3,
                do.inverseAlm = TRUE,
                do.normalizeAlm = TRUE,
                alphasExp = c(0.20, 0.50, 0.70, 0.95),
                weights = data.frame(myWeights = runif(20)))

}
\seealso{
\code{\link{get_hows}}
}
