% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{extract_peakdocs}
\alias{extract_peakdocs}
\title{Extract documents related to sentiment peaks}
\usage{
extract_peakdocs(sentomeasures, corpus, n = 10, type = "both",
  average = FALSE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}

\item{n}{a \code{numeric} value to indicate the number of dates to extract. This doesn't necessarily occurs to the
\code{n} most extreme sentiment values, given that those may occur on only one date for example.}

\item{type}{a \code{character} value, either \code{"pos"}, \code{"neg"} or \code{"both"}.}

\item{average}{xxx.}

\item{sentocorpus}{the \code{sentocorpus} object created with \code{\link{sento_corpus}}, used for the construction
of the input \code{sentomeasures}.}
}
\value{
A \code{list} with as elements \code{"dates"}, \code{"ids"} and \code{"documents"}.
}
\description{
This function will give you the dates and documents for which aggregated sentiment was most extreme (highest,
lowest or highest in absolute terms).
}
\examples{
data("usnews")
data("lexicons")
data("valence")

# construct a sentomeasures object to start with
corpus <- sento_corpus(corpusdf = usnews)
corpusSample <- quanteda::corpus_sample(corpus, sample = 500)
l <- setup_lexicons(lexicons[c("LM_eng", "HENRY_eng")], valence[["valence_eng"]])
ctr <- ctr_agg(howTime = c("equal_weight", "linear"), by = "month", lag = 3)
sentomeasures <- sento_measures(corpusSample, l, ctr)

# extract the peaks
peaksAbs <- extract_peakdocs(sentomeasures, corpus, n = 5)
peaksPos <- extract_peakdocs(sentomeasures, corpus, n = 5, type = "pos")
peaksNeg <- extract_peakdocs(sentomeasures, corpus, n = 5, type = "neg")

}
\author{
Samuel Borms
}
