% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentmodel.R
\name{fill_measures}
\alias{fill_measures}
\title{Add and fill missing dates}
\usage{
fill_measures(sentmeasures, fillLatest = FALSE)
}
\arguments{
\item{sentmeasures}{a \code{sentmeasures} object.}

\item{fillLatest}{a logical, if \code{TRUE} fills added dates with most recent value.}
}
\value{
A modified \code{sentmeasures} object.
}
\description{
Adds missing dates between earliest and latest date, such that time series is continuous on a day-by-day basis. Fills in
these dates with either \code{NA} or the respective latest non-\code{NA} values.
}
