% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{fill_measures}
\alias{fill_measures}
\title{Add and fill missing dates}
\usage{
fill_measures(sentomeasures, fill = "zero")
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{fill}{an element of \code{c("zero", "latest")}, the first assumes missing dates represent zero sentoment, the
latter assumes missing dates represent constant sentiment.}
}
\value{
A modified \code{sentomeasures} object.
}
\description{
Adds missing dates between earliest and latest date, such that time series is continuous on a period-by-period
basis. Fills in these dates with either \code{0} or the respective latest non-missing value.
}
