% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{fill_measures}
\alias{fill_measures}
\title{Add and fill missing dates}
\usage{
fill_measures(sentomeasures, do.fillLatest = FALSE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{fillLatest}{a logical, if \code{TRUE} fills added dates with most recent value.}
}
\value{
A modified \code{sentomeasures} object.
}
\description{
Adds missing dates between earliest and latest date, such that time series is continuous on a period-by-period
basis. Fills in these dates with either \code{NA} or the respective latest non-\code{NA} values.
}
