% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{get_dates}
\alias{get_dates}
\title{Get the dates of the sentiment measures/time series}
\usage{
get_dates(sento_measures)
}
\arguments{
\item{sento_measures}{a \code{sento_measures} object created using \code{\link{sento_measures}}.}
}
\value{
The \code{"date"} column in \code{sento_measures[["measures"]]} as a \code{character} vector.
}
\description{
Returns the dates of the sentiment time series.
}
\author{
Samuel Borms
}
