% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hows}
\alias{get_hows}
\title{Supported options to perform aggregation into sentiment measures.}
\usage{
get_hows()
}
\value{
A list with the supported aggregation hows for words (within documents), docs (across documents, per date)
and time (across dates).
}
\description{
Call for information purposes only. Used within \code{ctrl_agg()} to check if supplied
aggregation hows are supported.
}
