% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hows}
\alias{get_hows}
\title{Options supported to perform aggregation into sentiment measures}
\usage{
get_hows()
}
\value{
A list with the supported aggregation hows for arguments \code{howWithin} (\code{"words"}), \code{howDows}
(\code{"docs"}) and \code{howTime} (\code{"time"}), to be supplied to \code{\link{ctr_agg}}.
}
\description{
Outputs the supported aggregation arguments. Call for information purposes only. Used within
\code{\link{ctr_agg}} to check if supplied aggregation hows are supported.
}
\details{
See the package's \href{https://ssrn.com/abstract=3067734}{vignette} for a thoughtful explanation of
the different aggregation options. The \code{howWithin = "proportionalPol"} option divides each document's
initial sentiment score by the number of polarized words found in the document (counting each word only once
would it appear multiple times), instead of the total number of words which the \code{"proportional"} option gives.
}
\seealso{
\code{\link{ctr_agg}}
}
