% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hows}
\alias{get_hows}
\title{Options supported to perform aggregation into sentiment measures}
\usage{
get_hows()
}
\value{
A list with the supported aggregation hows for arguments \code{howWithin} (within documents), \code{howDows}
(across documents, per date) and \code{howTime} (across dates), to be supplied to \code{\link{ctr_agg}}.
}
\description{
Call for information purposes only. Used within \code{\link{ctr_agg}} to check if supplied
aggregation hows are supported.
}
\seealso{
\code{\link{ctr_agg}}
}
