% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{merge_measures}
\alias{merge_measures}
\title{Merge sentiment measures}
\usage{
merge_measures(ctr)
}
\arguments{
\item{ctr}{output from a \code{ctr_merge()} call.}
}
\value{
A modified \code{sentomeasures} object, with only the sentiment measures required, including updated information
and statistics, but the original sentiment scores \code{data.table} untouched.
}
\description{
Merge (further aggregate) measures by combining across the lexicons, features and time weighting schemes
dimensions. The combination occurs by taking the mean of the relevant measures.
}
\seealso{
\code{\link{ctr_merge}}
}
