% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{perform_agg}
\alias{perform_agg}
\title{Aggregate textual sentiment across documents and time}
\usage{
perform_agg(toAgg, ctr)
}
\arguments{
\item{toAgg}{output from a \code{compute_sentiment()} call, a list with as main component a sentiment scores
\code{data.table} with dates and feature--lexicon sentiment scores columns.}

\item{ctr}{output from a \code{ctr_agg()} call.}
}
\value{
a \code{sentomeasures} object.
}
\description{
Condense document-level textual sentiment scores into a panel of textual sentiment
measures by aggregating across documents and time.
}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{ctr_agg}}
}
