% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{perform_agg}
\alias{perform_agg}
\title{Aggregate textual sentiment across documents and time}
\usage{
perform_agg(toAgg, ctr)
}
\arguments{
\item{toAgg}{output from a \code{compute_sentiment()} call, a list with as main component a sentiment scores
\code{data.table} with dates and lexicon--feature sentiment scores columns.}

\item{ctr}{output from a \code{ctr_agg()} call.}
}
\value{
A \code{sentomeasures} object.
}
\description{
Condense document-level textual sentiment scores into a panel of textual sentiment
measures by aggregating across documents and time.
}
\examples{
data("useconomynews")
data("lexicons")
data("valence")

# computation of sentiment and aggregation into sentiment measures
corpus <- sento_corpus(corpusdf = useconomynews)
l <- setup_lexicons(lexicons[c("LM_eng", "HENRY_eng")], valence[["valence_eng"]])
sent <- compute_sentiment(corpus, l, how = "counts")
ctr <- ctr_agg(howTime = c("linear"), by = "year", lag = 3)
sentomeasures <- perform_agg(sent, ctr)

}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{ctr_agg}}
}
\author{
Samuel Borms, Keven Bluteau
}
