% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{plot.sentomeasures}
\alias{plot.sentomeasures}
\title{Plot sentiment measures}
\usage{
\method{plot}{sentomeasures}(sentomeasures, ...)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{...}{currently not used.}
}
\value{
A simple ggplot2 plot.
}
\description{
Generic plotting function for all sentiment measures in the provided \code{sentomeasures} object in one
plot.
}
