% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomodel.R
\name{plot.sentomodeliter}
\alias{plot.sentomodeliter}
\title{Plot iterative forecasts versus realized values}
\usage{
\method{plot}{sentomodeliter}(x, ...)
}
\arguments{
\item{x}{a \code{sentomeasures} object.}

\item{...}{not used.}
}
\value{
Returns a simple \pkg{ggplot2} plot, which can be added onto (or to alter its default elements) by using the
\code{+} operator (see examples).
}
\description{
Displays a plot of all forecasts made through the iterative model computation as incorporated in the
input \code{sentomodeliter} object, as well as the corresponding true values.
}
\examples{
data("useconomynews")
data("lexicons")
data("valence")
data("epu")

# construct a sentomeasures object to start with
useconomynews <- useconomynews[date >= "1980-01-01" & date < "2014-10-01", ]
corpus <- sento_corpus(corpusdf = useconomynews)
l <- setup_lexicons(lexicons[c("LM_eng", "HENRY_eng")], valence[["valence_eng"]])
ctr <- ctr_agg(howWithin = "tf-idf", howDocs = "proportional",
               howTime = c("equal_weight", "linear", "almon"),
               by = "month", lag = 3, ordersAlm = 1:3,
               do.inverseAlm = TRUE, do.normalizeAlm = TRUE)
sentomeasures <- sento_measures(corpus, l, ctr)

# prepare y and other x variables
y <- epu[epu$date >= sentomeasures$measures$date[1], ]$index
length(y) == nrow(sentomeasures$measures) # TRUE
x <- data.frame(runif(length(y)), rnorm(length(y))) # two other (random) x variables
colnames(x) <- c("x1", "x2")

# estimate regression iteratively based on a sample of 120, skipping the first 275 samples
ctr <- ctr_model(model = "gaussian", type = "AIC", do.iter = TRUE,
                 h = 0, nSample = 120, start = 276)
out <- sento_model(sentomeasures, y, x = x, ctr = ctr)
summary(out)

# plotting
p <- plot(out)
p <- p +
  ggthemes::theme_few()
p

}
\author{
Samuel Borms
}
