% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{plot_attributions}
\alias{plot_attributions}
\title{Plot forecasting attribution at specified level}
\usage{
plot_attributions(attributions, group = "features")
}
\arguments{
\item{attributions}{an output from a \code{retrieve_attributions()} call.}

\item{group}{a value from \code{c("lexicons", "features", "time", "lags")}.}
}
\value{
Returns a simple \pkg{ggplot2} plot, which can be added onto (or to alter its default elements) by using the
\code{+} operator (see examples). By default, a legend is positioned at the top if the number of dimensions (thus, individual
plots) is at maximum twelve.
}
\description{
Shows a plot of the attributions along the dimension provided.
}
\details{
See \code{\link{sento_model}} for an elaborate modelling example including the calculation and plotting of attributions.
This function does not handle the plotting of the attribution of individual documents, since there are often a lot of documents
involved and de facto they appear only once at one date (even though a document may contribute to forecasts at several dates,
depending on the number of lags in the time aggregation.)
}
\author{
Samuel Borms
}
