% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{retrieve_attributions}
\alias{retrieve_attributions}
\title{Retrieve top-down sentiment attributions given forecasting model object}
\usage{
retrieve_attributions(model, sentomeasures, ...)
}
\arguments{
\item{model}{a \code{sentomodel} or \code{sentomodeliter} object.}

\item{sentomeasures}{the \code{sentomeasures} object used to estimate the \code{sentomodel} object argument.}

\item{...}{the dates at which attribution is to be performed. The dates should be between the latest date avaiable in the
input \code{sentomeasures} object and the first date of the sample used to estimate the model in the \code{model} argument.
If not provided, attribution will be calculated for all in-sample dates. Not used if \code{model} is a \code{sentomodeliter}
object, for which attribution for all out-of-sample forecasting dates is calculated.}
}
\value{
A list with all possible dimensions for which aggregation is computed, being \code{"documents"}, \code{"lexicons"},
\code{"features"}, \code{"time"} and \code{"lags"}. The last four dimensions are \code{data.table}s having a \code{"date"}
column and the other columns the different names of the dimension, with the attributions as values. For document-level attribution,
the list is further decomposed into a \code{data.table} per date, with \code{"id"}, \code{date} and \code{attrib} columns.
}
\description{
Computes the attributions to forecasts for a (given) number of dates at all possible sentiment dimensions, based on
the coefficients associated to each sentiment measure, as estimated in the provided model object.
}
\details{
See \code{\link{sento_model}} for an elaborate modelling example including the calculation and plotting of attributions.
}
\seealso{
\code{\link{sento_model}}
}
\author{
Samuel Borms, Keven Bluteau
}
