% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{scale.sentomeasures}
\alias{scale.sentomeasures}
\title{Scaling and centering of sentiment measures}
\usage{
\method{scale}{sentomeasures}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a \code{sentomeasures} object.}

\item{center}{a \code{logical}, see documentation for the generic \code{\link{scale}}.}

\item{scale}{a \code{logical}, see documentation for the generic \code{\link{scale}}.}
}
\value{
A modified \code{sentomeasures} object, with the measures replaced by the scaled measures as well as updated
statistics.
}
\description{
Scales and centers sentiment measures from a \code{sentomeasures} object, column-per-column. By default, the
measures are normalized. \code{NA}s are removed first.
}
\examples{
# construct a sentomeasures object to start with
data("useconomynews")
corpus <- sento_corpus(corpusdf = useconomynews)
l <- setup_lexicons(lexicons[c("LM_eng", "HENRY_eng")], valence[["valence_eng"]])
ctr <- ctr_agg(howTime = c("equal_weight", "linear"), by = "year", lag = 3)
sentomeasures <- sento_measures(corpus, l, ctr)

# scale sentimeant measures
scaled <- scale(sentomeasures)

}
