% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{select_measures}
\alias{select_measures}
\title{Select a subset of sentiment measures}
\usage{
select_measures(sentomeasures, toSelect, do.all = TRUE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object.}

\item{toSelect}{a vector of components (lexicon, time weighting and features) which form the measures selected.}

\item{do.all}{a \code{logical} indicating if only measures for wich all (\code{TRUE}) or at least one (\code{FALSE}) of the
selection components should occur in the subset.}
}
\value{
A modified \code{sentomeasures} object, with only the measures required.
}
\description{
Selects the subset of sentiment measures which include either all of the given selection components or at least
one of them.
}
