% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentoapp.R
\name{sento_app}
\alias{sento_app}
\title{Shiny application for sentiment computation and visualisation}
\usage{
sento_app()
}
\description{
A Shiny application to showcase many of the package's functionalities.
}
\details{
This Shiny application demonstrates mainly the \code{compute_sentiment}, \code{corpus_summarize} and
\code{aggregate.sentiment} functions to gain insights in an uploaded corpus. The corpus should be uploaded in
.csv format (by default the \code{\link{usnews}} dataset is displayed). Lexicons and valence shifters can
be chosen from the built-in options or uploaded, and the weighting schemes are those available in the package.
All calculated values and statistics can be downloaded as a .csv file also.
}
\author{
Jeroen Van Pelt, Samuel Borms
}
