% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{sento_measures}
\alias{sento_measures}
\title{One-way road towards sentomeasures object}
\usage{
sento_measures(corpuS, lexicons, ctr)
}
\arguments{
\item{corpuS}{a \code{corpuS} object.}

\item{lexicons}{output from a \code{setup_lexicons()} call.}

\item{ctr}{output from a \code{ctr_agg()} call.}
}
\value{
A \code{sentomeasures} object.
}
\description{
Wrapper function which assembles calls to \code{compute_sentiment()} and \code{perform_agg()}, and includes
the input \code{corpuS} and computed sentiment scores in its output. Serves as the most direct way towards a panel of
textual sentiment measures, and a \code{sentomeasures} object.
}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{perform_agg}}
}
