% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{sento_measures}
\alias{sento_measures}
\title{One-way road towards a sentomeasures object}
\usage{
sento_measures(sentocorpus, lexicons, ctr)
}
\arguments{
\item{sentocorpus}{a \code{sentocorpus} object.}

\item{lexicons}{output from a \code{setup_lexicons()} call.}

\item{ctr}{output from a \code{ctr_agg()} call.}
}
\value{
A \code{sentomeasures} object, which is a list containing:
\item{measures}{a \code{data.table} with a \code{date} column and all textual sentiment measures as remaining columns.}
\item{features}{a \code{character} vector of the different features.}
\item{lexicons}{a \code{character} vector of the different lexicons used.}
\item{time}{a \code{character} vector of the different time weighting schemes used.}
\item{by}{a single \code{character} vector specifying the time interval of aggregation used.}
\item{stats}{a \code{data.frame} with a series of elementary statistics (mean, standard deviation, maximum, minimum, and
average correlation with all other measures) for each individual sentiment measure.}
\item{sentiment}{a sentiment scores \code{data.table} with a \code{date} and feature--lexicon sentiment scores columns.}
\item{howWithin}{a \code{character} vector to remind how sentiment within documents was aggregated.}
\item{howDocs}{a \code{character} vector to remind how sentiment across documents was aggregated.}
}
\description{
Wrapper function which assembles calls to \code{compute_sentiment()} and \code{perform_agg()}, and includes
the input \code{sentocorpus} and computed sentiment scores in its output. Serves as the most direct way towards a panel of
textual sentiment measures, and a \code{sentomeasures} object.
}
\examples{
# computation of sentiment measures based on a provided control function
data("useconomynews")
corpus <- sento_corpus(corpusdf = useconomynews)
l <- setup_lexicons(lexicons[c("LM_eng", "HENRY_eng")], valence[["valence_eng"]])
ctr <- ctr_agg(howWithin = "tf-idf",
               howDocs = "proportional",
               howTime = c("equal_weight", "linear", "almon"),
               by = "month",
               lag = 3,
               ordersAlm = 1:3,
               do.inverseAlm = TRUE,
               do.normalizeAlm = TRUE)
sentomeasures <- sento_measures(corpus, l, ctr)

}
\seealso{
\code{\link{compute_sentiment}}, \code{\link{perform_agg}}
}
