% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sentometrics-deprecated}
\alias{sentometrics-deprecated}
\alias{fill_measures}
\alias{merge_measures}
\alias{to_global}
\alias{ctr_merge}
\alias{subset_measures}
\alias{select_measures}
\alias{extract_peakdocs}
\title{Deprecated functions}
\usage{
fill_measures(sentomeasures, fill = "zero")

merge_measures(...)

to_global(sentomeasures, lexicons, features, time)

ctr_merge(...)

subset_measures(sentomeasures, subset)

select_measures(sentomeasures, toSelect)

extract_peakdocs(sentomeasures, sentocorpus, n = 10, type = "both",
  do.average = FALSE)
}
\arguments{
\item{sentomeasures}{a \code{sentomeasures} object created using \code{\link{sento_measures}}.}

\item{fill}{an element of \code{c("zero", "latest", NA)}; the first and last assume missing dates represent zero sentiment,
the second assumes missing dates represent constant sentiment.}

\item{...}{(other) allowed input arguments.}

\item{lexicons}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$lexicons)}, in the same order.
By default set to 1, which means equally weighted.}

\item{features}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$features)}, in the same order.
By default set to 1, which means equally weighted.}

\item{time}{a \code{numeric} vector of weights, of size \code{length(sentomeasures$time)}, in the same order. By default
set to 1, which means equally weighted.}

\item{subset}{a logical expression indicating the rows to keep.}

\item{toSelect}{a \code{character} vector of the lexicon, feature and time weighting scheme names, to indicate which
measures need to be selected, or as a \code{list} of \code{character} vectors, possibly with separately specified
combinations (only consisting of one lexicon, one feature, and one time weighting scheme at maximum).}

\item{sentocorpus}{the \code{sentocorpus} object created with \code{\link{sento_corpus}}, used for the construction
of the input \code{sentomeasures} object.}

\item{n}{a \code{numeric} value to indicate the number of dates associated to sentiment peaks to extract.}

\item{type}{a \code{character} value, either \code{"pos"}, \code{"neg"} or \code{"both"}, respectively to look
for the \code{n} dates related to the most positive, most negative or most extreme (in absolute terms) sentiment
occurrences.}

\item{do.average}{a \code{logical} to indicate whether peaks should be selected based on the average sentiment
value per date.}
}
\description{
Functions deprecated due to changed naming or because functionality is discarded. Deprecated functions are made defunct
every 1 major or every 2 minor package updates. See the NEWS file for more information about since when or why functions
have been deprecated.
}
\seealso{
\code{\link{measures_fill}}

\code{\link{measures_merge}}

\code{\link{measures_global}}

\code{\link{measures_subset}}

\code{\link{measures_select}}

\code{\link{peakdocs}}
}
