% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures.R
\name{setup_lexicons}
\alias{setup_lexicons}
\title{Setup lexicons format}
\usage{
setup_lexicons(lexiconsRaw, valenceWords = NULL, do.split = FALSE)
}
\arguments{
\item{lexiconsRaw}{a list of lexicons, with each element being a \code{data.frame} with a words column and a polarity score
column. The lexicons should be appropriately named for clarity in terms of subsequently obtained sentiment measures.
Alternatively, this argument can be a \code{character} vector specifying which built-in lexicons (\code{Sentometrics::LEXICONS}) to
use.}

\item{valenceWords}{a \code{data.frame} with a words column and a type indicating ... ### (negators, amplifiers, etc.)
Alternatively, this argument can be a \code{character} vector specifying which built-in valence words to use.}

\item{do.split}{a \code{logical} that if \code{TRUE} splits every lexicon into a separate positive polarity and negative polarity
lexicon.}
}
\value{
a list with each lexicon as a \code{data.table} list element according to its name, and the element
\code{valenceWords} that comprises the valence words. Every \code{x} column contains the words, every \code{y} column
containts the polarity or other quantitative information.
}
\description{
Structures provided lexicons and potentially valence words. Makes use of the \code{as_key()} function from the
\pkg{quanteda} package.
}
\seealso{
\code{\link[sentimentr]{as_key}}
}
