% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentometrics.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{Monthly S&P 500 Index returns}
\format{...}
\source{
\href{https://finance.yahoo.com/quote/\%5EGSPC/history?p=\%5EGSPC}{S&P 500 (^GSPC) at Yahoo Finance}
}
\usage{
SP500
}
\description{
Monthly returns for the S&P 500 Index between March 1988 and December 2014.

\itemize{
  \item date. Date as \code{"yyyy-mm-01"}.
  \item return. A \code{numeric} monthly return value.
}
}
\keyword{datasets}
