% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentomeasures_methods.R
\name{subset.sento_measures}
\alias{subset.sento_measures}
\title{Subset sentiment measures}
\usage{
\method{subset}{sento_measures}(x, subset, select, delete, ...)
}
\arguments{
\item{x}{a \code{sento_measures} object created using \code{\link{sento_measures}}.}

\item{subset}{a logical expression indicating the rows to keep.}

\item{select}{...}

\item{delete}{...}

\item{...}{...}
}
\value{
A modified \code{sento_measures} object, with only the kept rows, including updated information
and statistics, but the original sentiment scores \code{data.table} untouched.
}
\description{
Subsets rows of the sentiment measures based on its columns.
}
\examples{
data("usnews", package = "sentometrics")
data("list_lexicons", package = "sentometrics")
data("list_valence_shifters", package = "sentometrics")

# construct a sento_measures object to start with
corpus <- sento_corpus(corpusdf = usnews)
corpusSample <- quanteda::corpus_sample(corpus, size = 500)
l <- sento_lexicons(list_lexicons[c("LM_en", "HENRY_en")], list_valence_shifters[["en"]])
ctr <- ctr_agg(howTime = c("equal_weight", "linear"), by = "year", lag = 3)
sento_measures <- sento_measures(corpusSample, l, ctr)

# different subsets
sub1 <- measures_subset(sento_measures, HENRY_en--economy--equal_weight >= 0.01)
sub2 <- measures_subset(sento_measures,
   date \%in\% seq(as.Date("2000-01-01"), as.Date("2013-12-01"), by = "month"))

}
\author{
Samuel Borms
}
