% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentometrics.R
\docType{data}
\name{VALENCE}
\alias{VALENCE}
\title{Built-in valence word lists}
\format{A list with all built-in valence word lists, appropriately named.}
\source{
\code{\link[lexicon]{hash_valence_shifters}} (NEGATORS)
}
\usage{
VALENCE
}
\description{
A list containing all built-in valence word lists, with two columns: a \code{x} column with the words, and a \code{y}
column with the value associated to that word and type of valence shifter. The list element names incorporate
consecutively the type of valence shifter, name and language, and \code{"_tr"} as suffix if the word list is translated.
The word lists are in the form required for further sentiment analysis. The built-in valence word lists are the following:

\itemize{
  \item NEGATORS_ENG
  \item NEGATORS_FR_tr
  \item NEGATORS_NL_tr
}
}
\keyword{datasets}
